requirements.txt
flask>=2.3.0
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
plotly>=5.14.0
openai>=1.2.0
google-cloud-bigquery>=3.9.0
db-dtypes>=1.1.0